Introduce Workspaces, Can now re-open Tasks, Basic Color Scheme
I made a toolbar and introduced a basic color scheme. On the toolbar,
the words "Work" or "Home" are clickable. Clicking it allows you to
switch to a new workspace. So now you can have a workspace for Home and
a separate one for Work.
Also, when Tasks are closed, you can click them in the closed list view
to re-open them (send them to Ongoing View).